+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/entities/Close.java
@@ -16,22 +16,23 @@
  *
  */
 
-package com.onlyoffice.integration.dto;
+package com.onlyoffice.integration.entities;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.onlyoffice.model.documenteditor.config.document.ReferenceData;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import javax.persistence.Entity;
+import javax.persistence.Table;
+
+@Entity
+@Table(name = "`close`")
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Reference {
-    private ReferenceData referenceData;
-    private String path;
-    private String link;
+public class Close extends AbstractEntity {
+    private String text;
+    private Boolean visible;
 }