+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/entities/User.java
@@ -48,4 +48,5 @@ public class User extends AbstractEntity {
     private Boolean avatar;
     private String image;
     private Goback goback;
+    private Close close;
 }