+++ web/documentserver-example/python/src/views/actions.py
@@ -333,6 +333,7 @@ def edit(request):
                 'submitForm': submitForm,  # if the Submit form button is displayed or not
                 # settings for the Open file location menu button and upper right corner button
                 'goback':  user.goback if user.goback is not None else '',
+                'close':  user.close if user.close is not None else '',
             }
         }
     }
@@ -380,6 +381,7 @@ def edit(request):
 
     context = {  # the data that will be passed to the template
         'cfg': json.dumps(edConfig),  # the document config in json format
+        'fileName': filename,
         'fileType': fileType,  # the file type of the document (text, spreadsheet or presentation)
         'apiUrl': config_manager.document_server_api_url().geturl(),  # the absolute URL to the api
         # the image which will be inserted into the document