+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Change Log
 
+- filename in editor page title
+- close editor
 - he-IL skin language
 - ConvertService.ashx address replaced with converter
 - coauthoring/CommandService.ashx address replaced with command
 - submitForm for all modes
-
 - golang: new integration example
 - golang: upload files to the server
 - golang: create blank files and files with sample content